fix scripts so they work on a fresh clone

+++ _build.sh
@@ -1,5 +1,15 @@
 #!/bin/bash
 set -e
+cd $(dirname ${BASH_SOURCE[0]})
+
+if [ ! -d _submodules/jekyll_alias_generator/.git ]; then
+  echo
+  echo "Submodules aren't initialized."
+  echo
+  echo "Run this command once to initialze the jekyll alias submodule."
+  echo "$ git submodule sync --init"
+  exit 1
+fi
 
 jekyll build
 rm _site/app.yaml
